fixed 2048_Game.py.
running this game was giving this error: Traceback (most recent call last):
  File "G:\เอกสาร\responsive-website\3.py", line 216, in <module>
    game2048 = Game( gamepanel)
               ^^^^^^^^^^^^^^^^
TypeError: Game() takes no arguments

it was due to doublescore not used in dunder method. so did a small changes:  Changed _init_ to __init__ in both classes. 
the game is not that exciting in next comit and pr I shall make a more playable game. May You have a fun and happy life for doing good work.